[frontendActions] Remove VcsPushDialogAdapterHolder
